LayoutTests/fast/dom/HTMLFormElement/named-property-getter-radionodelist-expected.txt

Tests that HTMLFormElement's named property getter returns a RadioNodeList if there are several matching elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS radioNodeList.__proto__ is RadioNodeList.prototype
PASS radioNodeList.length is 3
PASS radioNodeList[0] is radio1
PASS radioNodeList[1] is radio2
PASS radioNodeList[2] is radio3
PASS radioNodeList.value is "value2"
PASS successfullyParsed is true

TEST COMPLETE
